mapbox moved to Park component
